+++ Marketplace/MarketplaceWaiters.cs
@@ -54,5 +54,32 @@ namespace Oci.MarketplaceService
             );
             return new Waiter<GetPublicationRequest, GetPublicationResponse>(config, agent);
         }
+        /// <summary>
+        /// Creates a waiter using default wait configuration.
+        /// </summary>
+        /// <param name="request">Request to send.</param>
+        /// <param name="targetStates">Desired resource states. If multiple states are provided then the waiter will return once the resource reaches any of the provided states</param>
+        /// <returns>a new Oci.common.Waiter instance</returns>
+        public Waiter<GetWorkRequestRequest, GetWorkRequestResponse> ForWorkRequest(GetWorkRequestRequest request, params OperationStatus[] targetStates)
+        {
+            return this.ForWorkRequest(request, WaiterConfiguration.DefaultWaiterConfiguration, targetStates);
+        }
+
+        /// <summary>
+        /// Creates a waiter using the provided configuration.
+        /// </summary>
+        /// <param name="request">Request to send.</param>
+        /// <param name="config">Wait Configuration</param>
+        /// <param name="targetStates">Desired resource states. If multiple states are provided then the waiter will return once the resource reaches any of the provided states</param>
+        /// <returns>a new Oci.common.Waiter instance</returns>
+        public Waiter<GetWorkRequestRequest, GetWorkRequestResponse> ForWorkRequest(GetWorkRequestRequest request, WaiterConfiguration config, params OperationStatus[] targetStates)
+        {
+            var agent = new WaiterAgent<GetWorkRequestRequest, GetWorkRequestResponse>(
+                request,
+                request => client.GetWorkRequest(request),
+                response => targetStates.Contains(response.WorkRequest.Status.Value)
+            );
+            return new Waiter<GetWorkRequestRequest, GetWorkRequestResponse>(config, agent);
+        }
     }
 }